Add accessibility-tree and VLM locator MCP tools
Split tools.py into a tools/ package (_base, _handlers, _factories)
to keep every module under the 750-line limit, and add five new
semantic-locator tools: ac_a11y_list / ac_a11y_find / ac_a11y_click
target widgets through the accessibility tree (stable across visual
restyles), and ac_vlm_locate / ac_vlm_click fall back to a vision
language model for ad-hoc descriptions. Both paths are far more
robust than pixel-template matching for dynamic UIs.
